requirements.txt
duckdb>=1.0.0
python-dotenv>=1.0.1
flask>=3.0.0
flask-limiter>=3.6.0
requests>=2.31.0
gunicorn>=21.2.0

# PostgreSQL support (for high-concurrency production workloads)
psycopg2-binary>=2.9.9
